ai/requirements.txt
Flask==2.2.3
PyJWT=2.6.0
openai==0.27.4
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
flask-cors==3.0.10
logging==3.11.3
bson==0.5.10